third_party/blink/web_tests/platform/generic/http/tests/security/mixedContent/filesystem-url-in-iframe-expected.txt

ALERT: PASS (1/2): File written
This tests that filesystem URLs created in a secure context are treated as secure origins. This test passes if the iframe renders PASS correctly, and no console warning appears.

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS (2/2): File displayed
